backend/requirements.txt
fastapi
uvicorn
PyGithub
openai
python-dotenv
github_utils